+++ travis-install.sh
@@ -2,5 +2,4 @@
 pip install pytest-cov==2.3.1
 if [[ $1 == 3* ]]; then
     pip install aiohttp
-    pip install pytest-aiohttp
 fi