+++ covador/__init__.py
@@ -1,6 +1,6 @@
 from .utils import wrap_in, ValidationDecorator, dpass
 from .types import *
 
-version = '0.9.2'
+version = '0.9.5'
 schema = make_schema(Map)
 list_schema = make_schema(ListMap)