+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ dev
 
 * [Feature] `dpass` decorator helper allows use complex expression inline.
 
+
 0.9.1
 =====